noahdarveau/more treeshaking updates (#2574)
* app, chat, and geoLocation treeshaking refactoring
