+++ change/@microsoft-teams-js-5dd2d28d-3280-4f21-9ef4-572a6de474f6.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Made the `app` (now fully), `chat`, and `geoLocation` files treeshakable",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
