+++ packages/teams-js/test/private/privateAPIs.spec.ts
@@ -7,7 +7,7 @@ import {
   sendCustomEvent,
   sendCustomMessage,
 } from '../../src/private/privateAPIs';
-import * as app from '../../src/public/app';
+import * as app from '../../src/public/app/app';
 import { FrameContexts, HostClientType, HostName, TeamType } from '../../src/public/constants';
 import { Context, FileOpenPreference } from '../../src/public/interfaces';
 import { Utils } from '../utils';
