+++ packages/teams-js/test/public/liveShareHost.spec.ts
@@ -1,5 +1,5 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
-import * as app from '../../src/public/app';
+import * as app from '../../src/public/app/app';
 import { liveShare, LiveShareHost } from '../../src/public/liveShareHost';
 import { setUnitializedRuntime } from '../../src/public/runtime';
 import { Utils } from '../utils';
