+++ packages/teams-js/test/public/publicAPIs.spec.ts
@@ -1,6 +1,6 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import * as utilFunc from '../../src/internal/utils';
-import * as app from '../../src/public/app';
+import * as app from '../../src/public/app/app';
 import { HostClientType, TeamType, UserTeamRole } from '../../src/public/constants';
 import { FrameContexts } from '../../src/public/constants';
 import { Context, FrameContext, LoadContext, TabInstanceParameters } from '../../src/public/interfaces';
