+++ packages/teams-js/test/public/search.spec.ts
@@ -1,7 +1,7 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import { GlobalVars } from '../../src/internal/globalVars';
 import { DOMMessageEvent } from '../../src/internal/interfaces';
-import * as app from '../../src/public/app';
+import * as app from '../../src/public/app/app';
 import { FrameContexts } from '../../src/public/constants';
 import { search } from '../../src/public/search';
 import { Utils } from '../utils';
